requirements.txt
# Core dependencies
pymisp>=2.4.187
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0

# Optional: for structured logging
# python-json-logger>=2.0.7
